Refactor x 1million to pull team stats into team manager and pass tests
