+++ lib/game.rb
@@ -6,12 +6,9 @@ class Game
               :away_team_id,
               :home_team_id,
               :away_goals,
-              :home_goals,
-              :manager,
-              :game_data
+              :home_goals
 
   def initialize(game_data, manager)
-    @manager      = manager
     @game_id      = game_data[:game_id].to_i
     @season       = game_data[:season].to_s
     @type         = game_data[:type]
@@ -20,6 +17,5 @@ class Game
     @home_team_id = game_data[:home_team_id]
     @away_goals   = game_data[:away_goals].to_i
     @home_goals   = game_data[:home_goals].to_i
-    @game_data    = game_data
   end
 end
