+++ lib/team.rb
@@ -1,19 +1,17 @@
 class Team
-  attr_reader :manager,
-              :team_id,
-              :team_data,
+  attr_reader :team_id,
+              :franchiseid,
               :teamname,
               :abbreviation,
               :link
 
   def initialize(team_data, manager)
-    @manager      = manager
-    @team_data    = team_data
     @team_id      = team_data[:team_id]
+    @franchiseid  = team_data[:franchiseid]
     @teamname     = team_data[:teamname]
     @abbreviation = team_data[:abbreviation]
     @link         = team_data[:link]
   end
 
-  
+
 end
